relocate tracer client run logic as client's method
